+++ packages/jest-config/src/__tests__/normalize.test.js
@@ -1251,9 +1251,11 @@ describe('preset with globals', () => {
       '/node_modules/global-foo/jest-preset.json',
       () => ({
         globals: {
+          __DEV__: false,
           config: {
             hereToStay: 'This should stay here',
           },
+          myString: 'hello world',
         },
       }),
       {virtual: true},
@@ -1268,9 +1270,11 @@ describe('preset with globals', () => {
     const {options} = normalize(
       {
         globals: {
+          __DEV__: true,
           config: {
             sideBySide: 'This should also live another day',
           },
+          myString: 'hello sunshine',
           textValue: 'This is just text',
         },
         preset: 'global-foo',
@@ -1279,17 +1283,15 @@ describe('preset with globals', () => {
       {},
     );
 
-    expect(options).toEqual(
-      expect.objectContaining({
-        globals: {
-          config: {
-            hereToStay: 'This should stay here',
-            sideBySide: 'This should also live another day',
-          },
-          textValue: 'This is just text',
-        },
-      }),
-    );
+    expect(options.globals).toEqual({
+      __DEV__: true,
+      config: {
+        hereToStay: 'This should stay here',
+        sideBySide: 'This should also live another day',
+      },
+      myString: 'hello sunshine',
+      textValue: 'This is just text',
+    });
   });
 });